Personal Assistant/requirements.txt
requests>=2.31.0
google-cloud-billing>=1.12.0
google-cloud-monitoring>=2.16.0
google-cloud-core>=2.4.1
